Replace minigrep args clone with iterator
- clone is inefficient because it allocates extra memory for the Config
instance to return with ownership of the Strings that are passed
- instead of collecting the args iterator and passing the strings, we
can pass the iterator directly along with its ownership. No need to
borrow a String slice
- args must be mutable because we will be iterating and modifying its
internal state
- restrict args parameter with trait bound generic type of an Iterator
with a String as the Item. i.e. any type that implements Iterator and
returns String items.
- Iterators return Options, so modify the variables to use match and
arms to retrieve values from the iterator next function